Move CharVocab to common, since it gets used in quite a few models
