+++ stanza/models/ner/vocab.py
@@ -1,9 +1,9 @@
 from collections import Counter, OrderedDict
 
-from stanza.models.common.vocab import BaseVocab, BaseMultiVocab
+from stanza.models.common.vocab import BaseVocab, BaseMultiVocab, CharVocab
 from stanza.models.common.vocab import VOCAB_PREFIX
 from stanza.models.common.pretrain import PretrainedWordVocab
-from stanza.models.pos.vocab import CharVocab, WordVocab
+from stanza.models.pos.vocab import WordVocab
 
 class TagVocab(BaseVocab):
     """ A vocab for the output tag sequence. """
